editted casting of s in sgn().

+++ NearNumber/BigInteger.h
@@ -99,7 +99,7 @@ public:
 
 	/* Returns arithmetic sign: -1 if negative, 1 if positive, and 0 otherwise. */
 	const intmax_t sgn(void) const {
-		return (1 - intmax_t(2) * (int)s) * operator bool();
+		return (1 - 2 * (intmax_t)s) * operator bool();
 	}
 
 	/* "xEZ" means "x is an arbitrary integer."
